test: trakt_models 并入 trakt_sync_service
- app.services.trakt.models 7 例并入 tests/services/test_trakt_sync_service.py
- 删除 test_trakt_models.py；api 路由层/models 层测试各自保留
